MNT: Warn if re-using the callback object
The current theory is that it is better to use a RunRouter to make new
instances per run rather than to feed more than one run through the
same callback instance.

The main motivation is to handle the case where more than run may
simultaneously be streaming into the system.   Using this method we
can put all of the logic for handling the multiplexing and routing in
one place (the RunRouter) rather than in every callback.